kos_zbot/requirements.txt
pyserial==3.5
grpcio
pykos>=0.7.10
adafruit-circuitpython-bno055
psutil
tabulate
tqdm
numpy
dotenv
pyaudio
sounddevice
pydub
pyee
openai[realtime]
protobuf==5.29.0 
click
rich
RPi.GPIO 
kscale>=0.3.16
kinfer